fixup! feat(builtin-sensors): introduce an STTS22H sensor driver

+++ src/sensors/ariel-os-sensor-stts22h/src/lib.rs
@@ -143,7 +143,7 @@ impl Stts22hI2c {
             }
 
             // Smaller text size than using `i32::from_be_bytes()`
-            let temp = i32::from(buf[1]) << 8 | i32::from(buf[0]);
+            let temp = i32::from(i16::from(buf[1]) << 8 | i16::from(buf[0]));
 
             let accuracy = accuracy(temp);
             let sample = Sample::new(temp, accuracy);
